requirements.txt









streamlit
pandas
numpy
scikit-learn
plotly
textblob
streamlit-lottie
qrcode
opencv-python
Pillow
pyzbar
python-dateutil







